+++ services/galley/test/integration/API.hs
@@ -28,7 +28,7 @@ where
 import API.CustomBackend qualified as CustomBackend
 import API.Federation qualified as Federation
 import API.MLS qualified
-import API.MLS.Util (convV8ToV9, memberToOtherMember)
+import API.MLS.Util (convV9ToV10, memberToOtherMember)
 import API.MessageTimer qualified as MessageTimer
 import API.Roles qualified as Roles
 import API.SQS
@@ -291,7 +291,7 @@ testGetConvQualifiedV2 = do
     fmap (unVersioned @'V2) . responseJsonError
       =<< getConvQualifiedV2 alice qcnv
         <!! const 200 === statusCode
-  liftIO $ conv @=? convV8ToV9 conv'
+  liftIO $ conv @=? convV9ToV10 conv'
 
 postProteusConvOk :: TestM ()
 postProteusConvOk = do
@@ -1511,7 +1511,7 @@ getConvsOk2 = do
   convs <- getAllConvs alice
   let c1 = find ((== cnvQualifiedId cnv1) . cnvQualifiedId) convs
   let c2 = find ((== cnv2.qualifiedId) . cnvQualifiedId) convs
-  liftIO . forM_ [(convV8ToV9 cnv1, c1), (cnv2, c2)] $ \(expected, mActual) ->
+  liftIO . forM_ [(convV9ToV10 cnv1, c1), (cnv2, c2)] $ \(expected, mActual) ->
     case mActual of
       Nothing -> assertFailure $ "Did not find expected conversation: " <> show expected
       Just actual -> do
@@ -1940,8 +1940,8 @@ postSelfConvOk = do
   let alice = qUnqualified qalice
   m <- postSelfConv alice <!! const 200 === statusCode
   n <- postSelfConv alice <!! const 200 === statusCode
-  mId <- assertConvV8 m SelfConv (Just alice) qalice [] Nothing Nothing
-  nId <- assertConvV8 n SelfConv (Just alice) qalice [] Nothing Nothing
+  mId <- assertConvV9 m SelfConv (Just alice) qalice [] Nothing Nothing
+  nId <- assertConvV9 n SelfConv (Just alice) qalice [] Nothing Nothing
   liftIO $ mId @=? nId
 
 postO2OConvOk :: TestM ()
@@ -1951,8 +1951,8 @@ postO2OConvOk = do
   connectUsers alice (singleton bob)
   a <- postO2OConv alice bob Nothing <!! const 200 === statusCode
   c <- postO2OConv alice bob Nothing <!! const 200 === statusCode
-  aId <- assertConvV8 a One2OneConv (Just alice) qalice [qbob] Nothing Nothing
-  cId <- assertConvV8 c One2OneConv (Just alice) qalice [qbob] Nothing Nothing
+  aId <- assertConvV9 a One2OneConv (Just alice) qalice [qbob] Nothing Nothing
+  cId <- assertConvV9 c One2OneConv (Just alice) qalice [qbob] Nothing Nothing
   liftIO $ aId @=? cId
 
 postConvO2OFailWithSelf :: TestM ()
@@ -1990,16 +1990,16 @@ postConnectConvOk = do
   n <-
     postConnectConv alice bob "Alice" "connect with me!" Nothing
       <!! const 200 === statusCode
-  mId <- assertConvV8 m ConnectConv (Just alice) qalice [] (Just "Alice") Nothing
-  nId <- assertConvV8 n ConnectConv (Just alice) qalice [] (Just "Alice") Nothing
+  mId <- assertConvV9 m ConnectConv (Just alice) qalice [] (Just "Alice") Nothing
+  nId <- assertConvV9 n ConnectConv (Just alice) qalice [] (Just "Alice") Nothing
   liftIO $ mId @=? nId
 
 postConnectConvOk2 :: TestM ()
 postConnectConvOk2 = do
   alice <- randomUser
   bob <- randomUser
-  m <- decodeConvIdV8 <$> req alice bob
-  n <- decodeConvIdV8 <$> req alice bob
+  m <- decodeConvIdV9 <$> req alice bob
+  n <- decodeConvIdV9 <$> req alice bob
   liftIO $ m @=? n
   where
     req alice bob =
@@ -2009,7 +2009,7 @@ putConvAcceptOk :: TestM ()
 putConvAcceptOk = do
   alice <- randomUser
   bob <- randomUser
-  qcnv <- decodeQualifiedConvIdV8 <$> postConnectConv alice bob "Alice" "come to zeta!" Nothing
+  qcnv <- decodeQualifiedConvIdV9 <$> postConnectConv alice bob "Alice" "come to zeta!" Nothing
   putConvAccept bob (qUnqualified qcnv) !!! const 200 === statusCode
   getConvQualified alice qcnv !!! do
     const 200 === statusCode
@@ -2023,7 +2023,7 @@ putConvAcceptRetry = do
   alice <- randomUser
   bob <- randomUser
   connectUsers alice (singleton bob)
-  cnv <- decodeConvIdV8 <$> postO2OConv alice bob (Just "chat")
+  cnv <- decodeConvIdV9 <$> postO2OConv alice bob (Just "chat")
   -- If the conversation type is already One2One, everything is 200 OK
   putConvAccept bob cnv !!! const 200 === statusCode
 
@@ -2036,13 +2036,13 @@ postMutualConnectConvOk = do
   ac <-
     postConnectConv alice bob "A" "a" Nothing
       <!! const 201 === statusCode
-  acId <- assertConvV8 ac ConnectConv (Just alice) qalice [] (Just "A") Nothing
+  acId <- assertConvV9 ac ConnectConv (Just alice) qalice [] (Just "A") Nothing
   bc <-
     postConnectConv bob alice "B" "b" Nothing
       <!! const 200 === statusCode
   -- The connect conversation was simply accepted, thus the
   -- conversation name and message sent in Bob's request ignored.
-  bcId <- assertConvV8 bc One2OneConv (Just alice) qbob [qalice] (Just "A") Nothing
+  bcId <- assertConvV9 bc One2OneConv (Just alice) qbob [qalice] (Just "A") Nothing
   liftIO $ acId @=? bcId
 
 postRepeatConnectConvCancel :: TestM ()
@@ -2245,7 +2245,7 @@ testGetQualifiedLocalConv :: TestM ()
 testGetQualifiedLocalConv = do
   alice <- randomUser
   convId <- decodeQualifiedConvId <$> postConv alice [] (Just "gossip") [] Nothing Nothing
-  conv :: ConversationV8 <- fmap responseJsonUnsafe $ getConvQualified alice convId <!! const 200 === statusCode
+  conv :: ConversationV9 <- fmap responseJsonUnsafe $ getConvQualified alice convId <!! const 200 === statusCode
   liftIO $ do
     assertEqual "conversation id" convId (C.cnvQualifiedId conv)
     assertEqual "conversation name" (Just "gossip") (C.cnvName conv)
@@ -2388,10 +2388,10 @@ testBulkGetQualifiedConvs = do
     let expectedFound =
           sortOn
             (.qualifiedId)
-            $ convV8ToV9 (remoteConversationView lAlice defMemberStatus (toRemoteUnsafe remoteDomainA mockConversationA))
-              : convV8ToV9 (remoteConversationView lAlice defMemberStatus (toRemoteUnsafe remoteDomainB mockConversationB))
+            $ convV9ToV10 (remoteConversationView lAlice defMemberStatus (toRemoteUnsafe remoteDomainA mockConversationA))
+              : convV9ToV10 (remoteConversationView lAlice defMemberStatus (toRemoteUnsafe remoteDomainB mockConversationB))
               : [localConv]
-        actualFound = sortOn (.qualifiedId) . map convV8ToV9 $ crFound convs
+        actualFound = sortOn (.qualifiedId) . map convV9ToV10 $ crFound convs
     assertEqual "found conversations" expectedFound actualFound
 
     -- Assumes only one request is made
@@ -2715,7 +2715,7 @@ removeRemoteMemberConvQualifiedFail = do
 deleteMembersQualifiedFailSelf :: TestM ()
 deleteMembersQualifiedFailSelf = do
   (alice, qalice) <- randomUserTuple
-  self <- decodeConvIdV8 <$> postSelfConv alice
+  self <- decodeConvIdV9 <$> postSelfConv alice
   qself <- Qualified self <$> viewFederationDomain
   deleteMemberQualified alice qalice qself !!! const 403 === statusCode
 
@@ -2724,7 +2724,7 @@ deleteMembersQualifiedFailO2O = do
   alice <- randomUser
   (bob, qbob) <- randomUserTuple
   connectUsers alice (singleton bob)
-  o2o <- decodeConvIdV8 <$> postO2OConv alice bob (Just "foo")
+  o2o <- decodeConvIdV9 <$> postO2OConv alice bob (Just "foo")
   qo2o <- Qualified o2o <$> viewFederationDomain
   deleteMemberQualified alice qbob qo2o !!! const 403 === statusCode
 
@@ -2745,7 +2745,7 @@ putQualifiedConvRenameOk = do
   qbob <- randomQualifiedUser
   let bob = qUnqualified qbob
   connectUsers alice (singleton bob)
-  conv <- decodeConvIdV8 <$> postO2OConv alice bob (Just "gossip")
+  conv <- decodeConvIdV9 <$> postO2OConv alice bob (Just "gossip")
   let qconv = Qualified conv (qDomain qbob)
   WS.bracketR2 c alice bob $ \(wsA, wsB) -> do
     void $ putQualifiedConversationName bob qconv "gossip++" !!! const 200 === statusCode
@@ -2764,7 +2764,7 @@ putConvDeprecatedRenameOk = do
   qbob <- randomQualifiedUser
   let bob = qUnqualified qbob
   connectUsers alice (singleton bob)
-  conv <- decodeConvIdV8 <$> postO2OConv alice bob (Just "gossip")
+  conv <- decodeConvIdV9 <$> postO2OConv alice bob (Just "gossip")
   let qconv = Qualified conv (qDomain qbob)
   WS.bracketR2 c alice bob $ \(wsA, wsB) -> do
     unversionedGalley <- view tsUnversionedGalley
@@ -2794,7 +2794,7 @@ putConvRenameOk = do
   qbob <- randomQualifiedUser
   let bob = qUnqualified qbob
   connectUsers alice (singleton bob)
-  conv <- decodeConvIdV8 <$> postO2OConv alice bob (Just "gossip")
+  conv <- decodeConvIdV9 <$> postO2OConv alice bob (Just "gossip")
   let qconv = Qualified conv (qDomain qbob)
   WS.bracketR2 c alice bob $ \(wsA, wsB) -> do
     void $ putConversationName bob conv "gossip++" !!! const 200 === statusCode
@@ -2937,7 +2937,7 @@ putMemberOk update = do
   qbob <- randomQualifiedUser
   let bob = qUnqualified qbob
   connectUsers alice (singleton bob)
-  conv <- decodeConvIdV8 <$> postO2OConv alice bob (Just "gossip")
+  conv <- decodeConvIdV9 <$> postO2OConv alice bob (Just "gossip")
   let qconv = Qualified conv (qDomain qbob)
   getConv alice conv !!! const 200 === statusCode
   -- Expected member state
@@ -3216,7 +3216,7 @@ postTypingIndicatorsV2 = do
 
   connectUsers alice (singleton bob)
 
-  conv <- decodeConvIdV8 <$> postO2OConv alice bob Nothing
+  conv <- decodeConvIdV9 <$> postO2OConv alice bob Nothing
   lcnv <- qualifyLocal conv
 
   WS.bracketR2 c alice bob $ \(wsAlice, wsBob) -> do
@@ -3262,7 +3262,7 @@ postTypingIndicators = do
 
   connectUsers alice (singleton bob)
 
-  conv <- decodeConvIdV8 <$> postO2OConv alice bob Nothing
+  conv <- decodeConvIdV9 <$> postO2OConv alice bob Nothing
   lcnv <- qualifyLocal conv
 
   WS.bracketR2 c alice bob $ \(wsAlice, wsBob) -> do
@@ -3305,7 +3305,7 @@ postTypingIndicatorsHandlesNonsense = do
   bob <- randomUser
 
   connectUsers alice (singleton bob)
-  conv <- decodeConvIdV8 <$> postO2OConv alice bob Nothing
+  conv <- decodeConvIdV9 <$> postO2OConv alice bob Nothing
 
   post
     ( g